Split expressions from general AST classes and simplify them
